Fixed: #1838 AS3 - Properly handling of long unsigned values, hex values, default uint values etc. Changed: AS3 integer values are internally (e.g. in the lib) handled as java int type instead of long.

+++ libsrc/ffdec_lib/src/com/jpexs/decompiler/flash/abc/avm2/instructions/stack/PushByteIns.java
@@ -12,7 +12,8 @@
  * Lesser General Public License for more details.
  * 
  * You should have received a copy of the GNU Lesser General Public
- * License along with this library. */
+ * License along with this library.
+ */
 package com.jpexs.decompiler.flash.abc.avm2.instructions.stack;
 
 import com.jpexs.decompiler.flash.abc.ABC;
@@ -45,7 +46,7 @@ public class PushByteIns extends InstructionDefinition implements PushIntegerTyp
 
     @Override
     public void translate(AVM2LocalData localData, TranslateStack stack, AVM2Instruction ins, List<GraphTargetItem> output, String path) {
-        stack.push(new IntegerValueAVM2Item(ins, localData.lineStartInstruction, (long) (byte) ins.operands[0]));
+        stack.push(new IntegerValueAVM2Item(ins, localData.lineStartInstruction, (int) (byte) ins.operands[0]));
     }
 
     @Override
